fix (error on deploy): fix params and type Wallet[]
+ (eslint.config.mjs) fix rules.
+ (page.tsx) fix error return data.
+ (route.ts wallet) fix params id
+ (columns.tsx and WalletActions.tsx) fix error useState().
+ (Navbar.tsx) fix navbar compononts none
+ tested deploy with "npm run build" all success
