+++ eslint.config.mjs
@@ -6,11 +6,17 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
 const compat = new FlatCompat({
-  baseDirectory: __dirname,
+    baseDirectory: __dirname,
 });
 
 const eslintConfig = [
-  ...compat.extends("next/core-web-vitals", "next/typescript"),
+    ...compat.config({
+        extends: ["next"],
+        rules: {
+            "react/no-unescaped-entities": "off",
+            "@next/next/no-page-custom-font": "off",
+        },
+    }),
 ];
 
 export default eslintConfig;
